+++ python/importer/run_locally.sh
@@ -18,4 +18,4 @@ fi
 
 # Now we can start the function.
 
-func start
+func start --port 7072
